% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marmot_function.R
\name{features}
\alias{features}
\title{Retrieve the complete feature list supported by marmot currently.}
\usage{
features(
  baseurl = httr::parse_url("http://marmot-proxy2-dev.us-east-2.elasticbeanstalk.com")
)
}
\arguments{
\item{baseurl}{The base URL of a marmot service.}
}
\value{
A list containing a name and a description property for each feature.
}
\description{
Retrieve the complete feature list supported by marmot currently.
}
\examples{
marmotR::features()
}
