% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marmotR-package.R
\docType{package}
\name{marmotR}
\alias{marmotR}
\alias{marmotR-package}
\title{marmotR: R connector package for the Marmot project}
\description{
This is the R connector package for the Marmot project.
}
\author{
\strong{Maintainer}: Anonymous Marmot \email{anonymous@marmot.com}

}
\keyword{internal}
