% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marmot_function.R
\name{search}
\alias{search}
\title{Search for matching packages.}
\usage{
search(
  query = "",
  limit = 50,
  baseurl = httr::parse_url("http://marmot-proxy2-dev.us-east-2.elasticbeanstalk.com")
)
}
\arguments{
\item{query}{The query in marmot query syntax.}

\item{limit}{The maximum number of results.}

\item{baseurl}{The base URL of a marmot service.}
}
\value{
A list containing the request data.
}
\description{
Search for matching packages.
}
\examples{
marmotR::search("[metrics.bytecode.ret]>0")
}
