% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marmot_function.R
\name{version}
\alias{version}
\title{Retrieve the version of a marmot service.}
\usage{
version(
  baseurl = httr::parse_url("http://marmot-proxy2-dev.us-east-2.elasticbeanstalk.com")
)
}
\arguments{
\item{baseurl}{The base URL of a marmot service.}
}
\value{
The version of the running service.
}
\description{
Retrieve the version of a marmot service.
}
\examples{
marmotR::version()
}
